remove datatable component
