+++ package.json
@@ -18,7 +18,7 @@
         "@dhis2/ui": "6.12.0",
         "@eyeseetea/d2-api": "1.14.0",
         "@eyeseetea/d2-logger": "1.0.0",
-        "@eyeseetea/d2-ui-components": "2.9.0-beta.1",
+        "@eyeseetea/d2-ui-components": "2.9.0-beta.2",
         "@eyeseetea/feedback-component": "0.0.3",
         "@material-ui/core": "4.12.4",
         "@material-ui/icons": "4.11.3",
@@ -28,8 +28,6 @@
         "d2": "31.10.2",
         "d2-manifest": "1.0.0",
         "font-awesome": "4.7.0",
-        "linkify-react": "^4.1.3",
-        "linkifyjs": "^4.1.3",
         "lodash": "4.17.21",
         "purify-ts": "2.0.3",
         "purify-ts-extra-codec": "0.6.0",
