+++ src/domain/usecases/common/UCIssue.ts
@@ -87,6 +87,7 @@ export class UCIssue {
                     id: undefined,
                     followUp: undefined,
                     step: undefined,
+                    search: undefined,
                 },
                 pagination: { page: 1, pageSize: 10 },
                 sorting: { field: "number", order: "asc" },
